Update getlambda.m
A piece of code was inserted at lines 40-46. This is done to remove any empty columns in DB_store and DF_store that can appear when the number of points queried (npt) on the test attractor is equal or close to the total number of points on the test attractor. Empty columns appeared because: for a point to be viable to track the separation of a trajectory between embd_lib and embd_test, then length(L) points are needed - however, length(L) future/past points are not available at either end of the embd_test and embd_lib trajectories, i.e. if embd_test(t=1) is chosen to track forward/backward trajectory separation through time, while the point embd_test(t=1+L) is available, the point embd_test(t =1-L) is not.